refactor(ci): only generate on main and cron run
When someone creates a pr it won't work to regenerate the index on their
branch since it won't be able to commit it back. So this splits CI into
two different jobs, one is just a check that will always run during prs
and the other runs during the merge and during the nightly cron to test,
build, and then commit the changes.